refactor mock
