--- frontend/app/.server/domain/exceptions/client-application-not-found.exception.ts
@@ -1,3 +1,0 @@
-import { ResourceNotFoundException } from './resource-not-found.exception';
-
-export class ClientApplicationNotFoundException extends ResourceNotFoundException {}
