+++ frontend/app/.server/domain/services/client-application.service.ts
@@ -1,6 +1,5 @@
 import { inject, injectable } from 'inversify';
 
-import { ClientApplicationNotFoundException } from '../exceptions/client-application-not-found.exception';
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
 import type { ClientApplicationDto } from '~/.server/domain/dtos';
 import type { ClientApplicationDtoMapper } from '~/.server/domain/mappers';
@@ -23,19 +22,17 @@ export interface ClientApplicationService {
    * Finds client application data by Social Insurance Number (SIN).
    *
    * @param sin The Social Insurance Number of the client.
-   * @returns A Promise that resolves to the client application data if found.
-   * @throws {ClientApplicationNotFoundException}
+   * @returns A Promise that resolves to the client application data if found, or `null` if not found.
    */
-  findClientApplicationBySin(sin: string): Promise<ClientApplicationDto>;
+  findClientApplicationBySin(sin: string): Promise<ClientApplicationDto | null>;
 
   /**
    * Finds client application data by first name, last name, date of birth, and client number.
    *
    * @param searchCriteria An object containing the search criteria.
-   * @returns A Promise that resolves to the client application data if found.
-   * @throws {ClientApplicationNotFoundException}
+   * @returns A Promise that resolves to the client application data if found, or `null` if not found.
    */
-  findClientApplicationByPersonalInfo(searchCriteria: FindByPersonalInfoSearchCriteria): Promise<ClientApplicationDto>;
+  findClientApplicationByPersonalInfo(searchCriteria: FindByPersonalInfoSearchCriteria): Promise<ClientApplicationDto | null>;
 }
 
 @injectable()
@@ -50,31 +47,19 @@ export class ClientApplicationServiceImpl implements ClientApplicationService {
     this.log = logFactory.createLogger('ClientApplicationServiceImpl');
   }
 
-  async findClientApplicationBySin(sin: string): Promise<ClientApplicationDto> {
+  async findClientApplicationBySin(sin: string): Promise<ClientApplicationDto | null> {
     this.log.debug('Get client application by sin');
     this.log.trace('Get client application with sin: [%s]', sin);
     const clientApplicationEntity = await this.ClientApplicationRepository.findClientApplicationBySin(sin);
-
-    if (!clientApplicationEntity) {
-      this.log.error('Client application with sin: [%s] not found', sin);
-      throw new ClientApplicationNotFoundException(`Client application with sin: [${sin}] not found`);
-    }
-
-    const clientApplicationDto = this.ClientApplicationDtoMapper.mapClientApplicationEntityToClientApplicationDto(clientApplicationEntity);
+    const clientApplicationDto = clientApplicationEntity ? this.ClientApplicationDtoMapper.mapClientApplicationEntityToClientApplicationDto(clientApplicationEntity) : null;
     this.log.trace('Returning client application: [%j]', clientApplicationDto);
     return clientApplicationDto;
   }
 
-  async findClientApplicationByPersonalInfo({ firstName, lastName, dateOfBirth, clientNumber }: FindByPersonalInfoSearchCriteria): Promise<ClientApplicationDto> {
+  async findClientApplicationByPersonalInfo({ firstName, lastName, dateOfBirth, clientNumber }: FindByPersonalInfoSearchCriteria): Promise<ClientApplicationDto | null> {
     this.log.debug('Get client application with first name: [%s], last name: [%s], date of birth: [%s], client number: [%s]', firstName, lastName, dateOfBirth, clientNumber);
     const clientApplicationEntity = await this.ClientApplicationRepository.findClientApplicationByCriteria({ firstName, lastName, dateOfBirth, clientNumber });
-
-    if (!clientApplicationEntity) {
-      this.log.error('Client application with first name: [%s], last name: [%s], date of birth: [%s], client number: [%s] not found', firstName, lastName, dateOfBirth, clientNumber);
-      throw new ClientApplicationNotFoundException(`Client friendly status with firstName: [${firstName}], lastName: [${lastName}], dateOfBirth: [${dateOfBirth}], clientNumber: [${clientNumber}] not found`);
-    }
-
-    const clientApplicationDto = this.ClientApplicationDtoMapper.mapClientApplicationEntityToClientApplicationDto(clientApplicationEntity);
+    const clientApplicationDto = clientApplicationEntity ? this.ClientApplicationDtoMapper.mapClientApplicationEntityToClientApplicationDto(clientApplicationEntity) : null;
     this.log.trace('Returning client application: [%j]', clientApplicationDto);
     return clientApplicationDto;
   }
